requirements.txt
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langchain-community
langchain-google-genai
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
numpy==2.0.2
openai==2.16.0
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tenacity==9.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
python-dotenv
google-search-results
google-generativeai
chromadb
pypdf
tiktoken
rank-bm25